backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16
channels>=4.1
channels-redis>=4.2
psycopg[binary]>=3.1
psycopg2-binary>=2.9.11; platform_system=="Windows"
reportlab>=4.2
celery>=5.4
djangorestframework-simplejwt>=5.3
firebase-admin>=6.5
